dev/back-end/requirements.txt
Flask==3.1.0
Flask_Cors==4.0.1
PyJWT==2.10.1
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1  
pymysql==1.1.1
bcrypt==4.2.1
cryptography==44.0.2
pillow==11.1.0
